LayoutTests/fast/dom/SelectorAPI/dumpNodeList-expected.txt

Line 1
Line 2
Line 3

line 4
line 5

Document.querySelectorAll

[object HTMLDivElement], length: 1
length: 0
[object HTMLDivElement], [object HTMLDivElement], length: 2
[object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 4
[object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 5
SyntaxError: '@font-face' is not a valid selector.
SyntaxError: '' is not a valid selector.
length: 0
length: 0
[object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], length: 4

Element.querySelectorAll

length: 0
[object HTMLDivElement], length: 1
length: 0
[object HTMLDivElement], [object HTMLDivElement], length: 2
[object HTMLDivElement], [object HTMLParagraphElement], length: 2
[object HTMLDivElement], [object HTMLParagraphElement], length: 2
SyntaxError: '@font-face' is not a valid selector.
SyntaxError: '' is not a valid selector.
length: 0
length: 0

DocumentFragment.querySelectorAll

[object HTMLDivElement], length: 1
[object HTMLDivElement], length: 1
length: 0
[object HTMLDivElement], [object HTMLDivElement], length: 2
[object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 3
[object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 3
SyntaxError: '@font-face' is not a valid selector.
SyntaxError: '' is not a valid selector.
length: 0
length: 0

Document.querySelector

[object HTMLDivElement]
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
SyntaxError: '@font-face' is not a valid selector.
SyntaxError: '' is not a valid selector.
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')

Element.querySelector

TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
[object HTMLDivElement]
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
SyntaxError: '@font-face' is not a valid selector.
SyntaxError: '' is not a valid selector.
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')

DocumentFragment.querySelector

[object HTMLDivElement]
[object HTMLDivElement]
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
SyntaxError: '@font-face' is not a valid selector.
SyntaxError: '' is not a valid selector.
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
TypeError: null is not an object (evaluating 'node.querySelector(selectorString).dump')
